refactor: slugifyanchor loc
